Add browserify support
add browserify module and split datalayer into modules
